requirements.txt
discord.py==2.4.1
python-dotenv==1.0.0